Work around bug by changing the limit
This wll work because the check is completely wrong. It's not looking
for a creference *within* 100 commits, it's looking for a reference
before the total number of commits exceeds 100, and even doing that
wrong because it ignores the latest build if the total number of
commits in it is more than 100.